requirements.txt
telethon
python-dotenv
pytz
filelock
